+++ Source/StageBuildDictionaryBase.swift
@@ -3,4 +3,4 @@ import Foundation
 /**
  Dictionary base mapping Semantic Version to Stages and Minimum Builds.
  */
-public typealias StageBuildDictionaryBase = [SemVer: [Stage: UInt8]]
+public typealias StageBuildDictionaryBase = [SemVer: [Stage: Int]]
